+++ test/integration/TestGas.js
@@ -1,0 +1,60 @@
+import { mineUntilPropose, mineUntilReveal } from '../helpers/SpecHelper';
+
+import registerVerifiers from '../helpers/RegisterVerifiers';
+import createProposals from '../samples/proposals';
+
+const ChainUtil = require('../proxy-contracts/proxyChain');
+
+const Chain = artifacts.require('Chain');
+
+const proxyChain = ChainUtil();
+
+contract('Chain - test GAS', (accounts) => {
+  let chainInstance;
+
+  const verifiersCount = 9;
+
+  const phaseDuration = verifiersCount * 5;
+
+  const proposalsObj = createProposals(verifiersCount, accounts);
+  const {
+    verifiersAddr, secrets, proposals, blindedProposals,
+  } = proposalsObj;
+
+
+  beforeEach(async () => {
+    const registryAddr = await registerVerifiers(accounts[0], verifiersAddr);
+
+    chainInstance = await Chain.new(
+      registryAddr,
+      phaseDuration,
+    );
+
+    proxyChain.setInstanceVar(chainInstance);
+    proxyChain.setFromVar(verifiersAddr[0]);
+
+    await mineUntilReveal(phaseDuration);
+  });
+
+
+  it('should be optimized', async () => {
+    // NOTICE: testing gas optimization - all below calls are perform only to check,
+    // how much gas it cost to do each trasactions,
+    // you can change the order or create another scenarios
+    // to see gar results, go to `truffle.js` and uncomment mocha - reporter, then run:
+    // truffle test <thisFile>
+    await mineUntilPropose(phaseDuration);
+
+    const awaits = [];
+    for (let i = 0; i < verifiersAddr.length; i += 1) {
+      awaits.push(chainInstance.propose(blindedProposals[i], { from: verifiersAddr[i] }));
+    }
+    await Promise.all(awaits);
+
+    await mineUntilReveal(phaseDuration);
+    await chainInstance.reveal(proposals[0], secrets[0], { from: verifiersAddr[0] });
+
+    await mineUntilPropose(phaseDuration);
+    await chainInstance.propose(blindedProposals[0], { from: verifiersAddr[0] });
+  });
+});