+++ test/proxy-utils/index.js
@@ -1,0 +1,5 @@
+import _proxyExecute from './proxyExecute';
+
+export const proxyExecute = _proxyExecute;
+
+export default proxyExecute;